refactor(pkg/trie/triedb): remove lookup struct
