--- pkg/trie/triedb/lookup_test.go
@@ -1,1 +1,0 @@
-package triedb
